+++ CHANGELOG.md
@@ -13,6 +13,10 @@
   [bootstraponline](https://github.com/bootstraponline)
   [#689](https://github.com/realm/SwiftLint/issues/689)
 
+* Add configuration for setting a warning threshold.  
+  [woodhamgh](https://github.com/woodhamgh)
+  [696](https://github.com/realm/SwiftLint/issues/696)
+
 ##### Bug Fixes
 
 * Fix LegacyConstructorRule when using variables instead of numbers.